% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_of_last_update.R
\name{date_of_last_update}
\alias{date_of_last_update}
\title{Get date of last municipal inventory update}
\usage{
date_of_last_update(mutations)
}
\arguments{
\item{mutations}{A tibble with municipality mutations (as created by
\code{\link{import_CH_municipality_inventory}})}
}
\value{
A \code{Date} object of length one.
}
\description{
Obtain the most recent change in the municipal inventory database. Most but
not all municipal mutations are made at first of january.
}
