% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_municipality_inventory.R
\name{download_municipality_inventory}
\alias{download_municipality_inventory}
\title{Download municipality inventory}
\usage{
download_municipality_inventory(
  url = "https://www.bfs.admin.ch/bfsstatic/dam/assets/11467405/master",
  path = getwd()
)
}
\arguments{
\item{url}{Character vector of length one. Link to the zip file containing the municipality inventory}

\item{path}{Character vector of length one. Destination of extracted xml file.}
}
\value{
Character vector of length one. File path to the extracted XML file.
}
\description{
This functions downloads and extracts the municipality inventory form a defined online source.
}
