% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date.R
\name{filter_date}
\alias{filter_date}
\title{Filter by date}
\usage{
filter_date(tbl, date)
}
\arguments{
\item{tbl}{A tibble}

\item{date}{A \code{POSIXct} object}
}
\value{
A tibble which is a subset of \code{tbl}
}
\description{
Filter for existing municipalities at a specific point in time.
}
