% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_url.R
\name{get_current_url}
\alias{get_current_url}
\title{Get URL of current XML data set}
\usage{
get_current_url()
}
\value{
Returns URL to municipality inventory
}
\description{
Extract the url from the static source web page.
}
