% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_changes.R
\name{most_recent_changes}
\alias{most_recent_changes}
\title{Most recent changes}
\usage{
most_recent_changes(mutations)
}
\arguments{
\item{mutations}{A tibble with municipality mutations (as created by
\code{\link{import_CH_municipality_inventory}})}
}
\description{
Returns a table with the most recently changed municipalities.
}
