% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmt-package.R
\docType{package}
\name{smmt}
\alias{SMMT}
\alias{SMMT-package}
\alias{smmt}
\title{SMMT - The Swiss Municipal Data Merger Tool Maps Municipalities Over Time}
\description{
In Switzerland, the landscape of municipalities is changing rapidly mainly
due to mergers. The Swiss Municipal Data Merger Tool automatically detects
these mutations and maps municipalities of an old state to municipalities of
a new state. This functionality is helpful when working with datasets that
are based on different spatial references. The spatial reference in this
context signifies a set of municipalities at a given point in time.
}
\details{
For detailed information and examples, see \code{\link{map_old_to_new_state}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ValValetl/SMMT}
  \item Report bugs at \url{https://github.com/ValValetl/SMMT/issues}
}

}
\author{
\strong{Maintainer}: Valentin Knechtl \email{valentinknechtl@gmail.com} (\href{https://orcid.org/0000-0002-8545-8883}{ORCID})

}
\keyword{internal}
