% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_municipality_inventory.R
\name{download_municipality_inventory}
\alias{download_municipality_inventory}
\title{Download municipality inventory}
\usage{
download_municipality_inventory(
  url = get_current_url(),
  path = getwd(),
  verbose = TRUE
)
}
\arguments{
\item{url}{Character vector of length one. Link to the zip file containing the municipality inventory}

\item{path}{Character vector of length one. Destination of extracted xml file.}

\item{verbose}{Get a message after download about the content of the inventory.}
}
\value{
Character vector of length one. File path to the extracted XML file.
}
\description{
This functions downloads and extracts the municipality inventory form a defined online source.
}
